Move into baseline_indexer folder, just to candidate vecs for now
