+++ CMakeLists.txt
@@ -44,4 +44,4 @@ add_subdirectory(dx2)
 add_subdirectory(h5read)
 add_subdirectory(baseline)
 add_subdirectory(spotfinder)
-add_subdirectory(indexer)
+add_subdirectory(baseline_indexer)
