+++ baseline_indexer/sites_to_vecs.cc
